+++ src/popper/utils/getBasePlacement.ts
@@ -1,4 +1,4 @@
-import { Placement, Side } from "@floating-ui/core/src/types";
+import type { Placement, Side } from "@floating-ui/core";
 
 export const getBasePlacement = (placement: Placement): Side => {
   return placement.split("-")[0] as Side;