+++ src/popper/utils/getCrossAxis.ts
@@ -1,4 +1,4 @@
-import { Axis } from "@floating-ui/core/src/types";
+import type { Axis } from "@floating-ui/core";
 
 export const getCrossAxis = (axis: Axis): Axis => {
   return axis === "x" ? "y" : "x";