+++ src/typings/obsidian-ex.d.ts
@@ -1,7 +1,7 @@
 import "obsidian";
 
 import { ToolBar } from "../modules/toolbar";
-import { MiniToolbarEvtName } from "../popper/define";
+import { SelectionToolbarEvtName } from "../popper/define";
 
 declare module "obsidian" {
   interface Menu {
@@ -14,7 +14,7 @@ declare module "obsidian" {
 
   interface Workspace {
     trigger(
-      name: typeof MiniToolbarEvtName,
+      name: typeof SelectionToolbarEvtName,
       toolbar: ToolBar,
       range: EditorRange,
       editor: Editor,